Разбросал контроллеры каждый в свой файл и сделал их реэкспорт. Модуль 4    Node.js
Разбросал контроллеры    addContactCtrls    getContactByIdCtrls   listContactsCtrls  removeContactCtrls    updateContactCtrls   updateStatusContactCtrls    из файла   contactsCtrls.js    в каждый свой соответствующий отдельный файл    добавив им нужные для них импорты и затем их сразу экспортнул. Затем сделал их реэкспорт. Для этого в файле
 index.js   импортнул    ctrlWrapper   и импортнул выше указанные контроллеры. И затем сразу экспортнул данные контроллеры, обернув их при этом при помощи    ctrlWrapper   В связи с выше описанными манипуляциями уже не было необходимости в файле   contacts.js    (который находится в папке    api   ) указывать полный путь для импорта контроллеров. Достаточно указать путь до папки, в которой находятся файлы с контроллерами. Также, в связи с выше описанными перестановками, удалил за ненадобностью файл    contactsCtrls.js   Модуль 4    Node.js
