+++ controllers/listContactsCtrls.js
@@ -1,0 +1,8 @@
+const { ContactMongoose } = require("../models");
+
+const listContactsCtrls = async (req, res) => {
+  const result = await ContactMongoose.find();
+  res.status(200).json(result);
+};
+
+module.exports = listContactsCtrls;
